Fix logging system - ready for testing

+++ robotcode/src/main/java/org/usfirst/frc/team1885/robot/autonomous/DriveStraightDistance.java
@@ -14,10 +14,10 @@ public class DriveStraightDistance extends Command{
 	
 	private final DriveTrain driveTrain;
 	private final NavX navx;
-	private final int distanceToTravel;
+	private final double distanceToTravel;
 	
-	private int initialLeftPosition;
-	private int initialRightPosition;
+	private double initialLeftPosition;
+	private double initialRightPosition;
 	
 	private double initialYaw;
 	
@@ -59,7 +59,7 @@ public class DriveStraightDistance extends Command{
 		return false;
 	}
 	
-	private int getAverageDistanceTravel(){
+	private double getAverageDistanceTravel(){
 		return (Math.abs(driveTrain.getLeftPosition() - initialLeftPosition) + 
 			   Math.abs(driveTrain.getRightPosition() - initialRightPosition))/2;
 	}
